backend_api/pyproject.toml
[project]
name = "backend-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.3",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "web3>=7.12.0",
]
